Fit: 201707 version of casa library, 4fd04f0965d27f188264080f9d7f3d70f934835b.
Remove rpfits:
Since rpfits seems not compatible with gfortran-7 so building it fail. So try disabling find_rpfits part.

Add NO_LINK to let pgplot with on testing in cmake, because somehow it does not find out the correct path of fortran which will be used to take building example of pgplot in cmake’s test. Not sure if homebrew suddenly update its fortran path/version shipped in its gcc.

Remove: QWT, DBus. Although
1. dbuspp-xml2cpp
2. In code/stdcasa/xerces.h, uses XMLCh
But they should be not important, and it should be safe to completely remove dbus when building.

Remove:
Java
Qt4
Rpfits

Use the below to remove some submodules which CARTA does not need:
https://safe.nrao.edu/wiki/pub/Software/CASA/CartaBuildInstructionsForUbuntu/imageanalysisonubuntu.diff